+++ src/app/(student)/question/[questionId]/page.module.css
@@ -6,20 +6,6 @@
   padding: 20px;
 }
 
-.title {
-  font-size: 3.5rem;
-  border-bottom: 1px solid #ccc;
-  padding-bottom: 2rem;
-}
-
-.chatContainer {
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-  overflow-y: auto;
-  padding-bottom: 100px; 
-}
-
 .inputContainer {
   position: fixed;
   bottom: 0;
