+++ src/components/organisms/MessageList/index.module.css
@@ -1,0 +1,13 @@
+.title {
+  font-size: 3.5rem;
+  border-bottom: 1px solid #ccc;
+  padding-bottom: 2rem;
+}
+
+.chatContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  overflow-y: auto;
+  padding-bottom: 100px; 
+}
